+++ Mailer/extensions/mailer-module/src/Repositories/JobQueueRepository.php
@@ -164,10 +164,17 @@ SQL;
         return $table->delete();
     }
 
-    public function deleteAllByEmail(string $email): int
+    /**
+     * @param array<string> $emails
+     */
+    public function deleteAllByEmails(array $emails): int
     {
+        if (count($emails) === 0) {
+            return 0;
+        }
+
         return $this->getTable()->where([
-            'email' => $email
+            'email' => $emails
         ])->delete();
     }
 }