+++ Mailer/extensions/mailer-module/src/Repositories/LogsRepository.php
@@ -29,7 +29,15 @@ class LogsRepository extends Repository
 
     public function allForEmail(string $email): Selection
     {
-        return $this->getTable()->where('email', $email);
+        return $this->allForEmails([$email]);
+    }
+
+    /**
+     * @param array<string> $emails
+     */
+    public function allForEmails(array $emails): Selection
+    {
+        return $this->getTable()->where('email', $emails);
     }
 
     public function add(string $email, string $subject, int $templateId, ?int $jobId = null, ?int $batchId = null, ?string $mailSenderId = null, ?int $attachmentSize = null, ?string $context = null): ActiveRow
@@ -55,10 +63,17 @@ class LogsRepository extends Repository
         ];
     }
 
-    public function deleteAllForEmail(string $email): int
+    /**
+     * @param array<string> $emails
+     */
+    public function deleteAllForEmails(array $emails): int
     {
+        if (count($emails) === 0) {
+            return 0;
+        }
+
         return $this->getTable()->where([
-            'email' => $email
+            'email' => $emails
         ])->delete();
     }